netcfg: Make NetCfgNotifSubscriptions use uint32_t as filter bit mask
Commit 93b6d3b7bc0e (netcfg: use proper C++ base type for
NetCfgChangeType) switched to use uint32_t as the base filter type.
But the NetCfgSubscriptions class was not updated accordingly, still
using uint16_t.

With this change, there is also no need to do any casting from
utin32_t to uint16_t.

+++ src/netcfg/netcfg-subscriptions.cpp
@@ -78,7 +78,7 @@ void NetCfgSubscriptions::Subscribe(const std::string &sender, uint32_t filter_f
     {
         throw NetCfgException("Invalid subscription flag, must be < 65535");
     }
-    subscriptions[sender] = static_cast<uint16_t>(filter_flags);
+    subscriptions[sender] = filter_flags;
     subscr_owners[sender] = creds_query->GetUID(sender);
 }
